backend/app/services/scrapers/quora_scraper.py
import os
import httpx
from bs4 import BeautifulSoup
from typing import List, Dict, Any, Optional
from dotenv import load_dotenv
from ...routers.research import ResearchResult

# Load environment variables
load_dotenv()

async def search(query: str, max_results: int = 50) -> List[ResearchResult]:
    """
    Search Quora for questions and answers matching the query.
    
    Args:
        query: The search query
        max_results: Maximum number of results to return
        
    Returns:
        List of ResearchResult objects
    """
    try:
        # In a real implementation, this would scrape Quora
        # For now, we'll return mock data
        
        # Example of how you would scrape Quora in a real implementation:
        """
        results = []
        
        # Format query for URL
        formatted_query = query.replace(" ", "+")
        url = f"https://www.quora.com/search?q={formatted_query}"
        
        # Make request
        async with httpx.AsyncClient() as client:
            response = await client.get(url)
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # Find question elements
            question_elements = soup.find_all('div', class_='question_container')
            
            for element in question_elements[:max_results]:
                question_title = element.find('span', class_='question_title').text
                question_url = "https://www.quora.com" + element.find('a')['href']
                
                # Get the top answer if available
                answer_element = element.find('div', class_='answer_content')
                answer_text = answer_element.text if answer_element else ""
                
                # Get author if available
                author_element = element.find('a', class_='user')
                author = author_element.text if author_element else "Anonymous"
                
                results.append(ResearchResult(
                    platform="quora",
                    source_url=question_url,
                    content=f"Q: {question_title}\n\nA: {answer_text}",
                    author=author,
                    date=None,  # Quora doesn't easily expose dates
                    engagement=None,  # Quora doesn't easily expose engagement metrics
                    sentiment=None  # Would be analyzed separately
                ))
        """
        
        # Mock data for demonstration
        results = [
            ResearchResult(
                platform="quora",
                source_url=f"https://www.quora.com/What-are-the-best-ways-to-solve-{query.replace(' ', '-')}",
                content=f"Q: What are the best ways to solve {query}?\n\nA: In my experience working with over 50 startups, the most effective approach is to start by clearly defining the problem. Many entrepreneurs rush to solutions without fully understanding the underlying issues...",
                author="Jane Smith, Startup Advisor",
                date=None,
                engagement=None,
                sentiment="positive"
            ),
            ResearchResult(
                platform="quora",
                source_url=f"https://www.quora.com/Why-is-{query.replace(' ', '-')}-so-difficult-for-startups",
                content=f"Q: Why is {query} so difficult for startups?\n\nA: The main reason startups struggle with this is lack of resources. Unlike established companies, startups don't have the luxury of dedicated teams to handle these challenges...",
                author="Michael Johnson, Serial Entrepreneur",
                date=None,
                engagement=None,
                sentiment="neutral"
            ),
            ResearchResult(
                platform="quora",
                source_url=f"https://www.quora.com/How-do-you-know-if-your-solution-to-{query.replace(' ', '-')}-is-working",
                content=f"Q: How do you know if your solution to {query} is working?\n\nA: The key metrics to track are customer retention and engagement. If users are sticking around and actively using your product, that's a strong signal...",
                author="Sarah Williams, Product Manager",
                date=None,
                engagement=None,
                sentiment="positive"
            )
        ]
        
        return results
        
    except Exception as e:
        print(f"Error searching Quora: {str(e)}")
        return [] 