backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tweepy==4.14.0
praw==7.7.1
supabase==1.0.3
openai==1.3.5
langchain==0.0.335
tiktoken==0.5.1
pytest==7.4.3